ADDED: Manivelas y salida funcional
Falta meter algún elemento de UI para mostrar el progreso de la manivela, pero por lo demás, todo funciona.
