+++ TheHallwaySol/Assets/lua/Scenes/scene1.lua
@@ -1,5 +1,5 @@
 ﻿scene1 = {
-    entities = {"defaultCamera", "sceneLight", "Maze"}
+    entities = {"defaultCamera", "sceneLight", "Maze", "Player"}
 }
 
 defaultCamera = {
@@ -8,7 +8,7 @@ defaultCamera = {
 
     Transform = {
         Active = true,
-        Position = {25, 750,25},
+        Position = {5.75 * 20, 300,5.75*20},
         Scale = {1,1,1},
         Rotation = {0,0,0}
     },
@@ -17,7 +17,7 @@ defaultCamera = {
         Active = true,
         Name = "MainCam",
         Background = {1,1,0},
-        LookAt = {25,0,25},
+        LookAt = {5.75*20,0,5.75*20},
         Width = 0,
         Height = 0,
         NearClipDistance = 1,
@@ -52,7 +52,7 @@ sceneLight = {
 
 Maze = {
     Active = true,
-    Components = {"Transform", "MazeCreator", "Lever", "Exit"},
+    Components = {"Transform", "MazeManager"},
 
     Transform = {
         Active = true,
@@ -60,16 +60,39 @@ Maze = {
         Scale = {1,1,1},
         Rotation = {0,0,0}
     },
-    MazeCreator = {
+    MazeManager = {
         Active = true,
         Size = 5,
-        Holes = 25,
-        Levers = 8
-    },
-    Lever = {
-
-    },
-    Exit = {
-
+        Holes = 200,
+        Levers = 1,
+        ChargeVel = 5,
+        UnchargeVel = 2,
     }
+}
+Player = 
+{
+    Active = true,
+    Components = {"Transform", "MeshRenderer", "PlayerMovement", "Rigidbody" },
+    
+    Transform = {
+        Active = true,
+        Position = {6.5,0,6.5},
+        Scale = {1,1,1},
+        Rotation = {0,0,0}
+    },
+    MeshRenderer = {
+        Mesh = "Suzanne.mesh"
+    },
+    PlayerMovement = {
+
+    },
+    Rigidbody = {
+        Type = "Hull",
+        Mass = 1,
+        Trigger = false,
+        Static = false,
+        PositionConstrains = {0,1,0},
+        RotationConstrains = {1,1,1}
+    },
+
 }
