+++ TheHallwaySol/Project/TheHallway/TheHallway.vcxproj
@@ -165,7 +165,7 @@
     <ClCompile Include="..\..\Src\TheHallway\Lever.cpp" />
     <ClCompile Include="..\..\Src\TheHallway\Lighter.cpp" />
     <ClCompile Include="..\..\Src\TheHallway\LoopMovement.cpp" />
-    <ClCompile Include="..\..\Src\TheHallway\MazeCreator.cpp" />
+    <ClCompile Include="..\..\Src\TheHallway\MazeManager.cpp" />
     <ClCompile Include="..\..\Src\TheHallway\PlayerMovement.cpp" />
     <ClCompile Include="..\..\Src\TheHallway\Source.cpp" />
   </ItemGroup>
@@ -174,7 +174,7 @@
     <ClInclude Include="..\..\Src\TheHallway\Lever.h" />
     <ClInclude Include="..\..\Src\TheHallway\Lighter.h" />
     <ClInclude Include="..\..\Src\TheHallway\LoopMovement.h" />
-    <ClInclude Include="..\..\Src\TheHallway\MazeCreator.h" />
+    <ClInclude Include="..\..\Src\TheHallway\MazeManager.h" />
     <ClInclude Include="..\..\Src\TheHallway\PlayerMovement.h" />
   </ItemGroup>
   <ItemGroup>
