+++ TheHallwaySol/Src/TheHallway/Exit.h
@@ -14,7 +14,7 @@ public:
     ~Exit() {}
 
     virtual bool init(luabridge::LuaRef parameterTable = { nullptr }) { return true; }
-    virtual void onCollisionEnter(QuackEntity* other, Vector3D point);
+    virtual void onTriggerEnter(QuackEntity* other, Vector3D point);
     static std::string GetName() { return "Exit"; }
 };
