+++ TheHallwaySol/Src/TheHallway/PlayerMovement.h
@@ -14,7 +14,7 @@ public:
 
 	virtual bool init(luabridge::LuaRef parameterTable = { nullptr }) override;
 
-	static std::string GetName() { return "PlaterMovement"; }
+	static std::string GetName() { return "PlayerMovement"; }
 
 	void move();
 	void update() override;
