+++ TheHallwaySol/Src/TheHallway/Source.cpp
@@ -2,7 +2,7 @@
 #include <fstream>
 #include "QuackEnginePro.h"
 #include "FactoryManager.h"
-#include "MazeCreator.h"
+#include "MazeManager.h"
 #include "Lighter.h"
 #include "PlayerMovement.h"
 #include "Lever.h"
@@ -33,7 +33,7 @@ WinMain(HINSTANCE zHInstance, HINSTANCE prevInstance, LPSTR lpCmdLine, int nCmdS
 
 	// AQUI FALTA MANEJO DE ERRORES Y EXCEPCIONES
 	if (QuackEnginePro::Init()) {
-		FactoryManager::instance()->add<MazeCreator>();
+		FactoryManager::instance()->add<MazeManager>();
 		FactoryManager::instance()->add<Lighter>();
 		FactoryManager::instance()->add<PlayerMovement>();
 		FactoryManager::instance()->add<Lever>();
